+++ src/typescript/Camera.ts
@@ -2,7 +2,6 @@ import * as THREE from 'three';
 import Time from './Utils/Time';
 import Sizes from './Utils/Sizes';
 import CameraControls from 'camera-controls';
-import Controls from './Controls';
 
 export default class Camera {
   // Utilities
@@ -20,8 +19,6 @@ export default class Camera {
   readonly debugFolder!: dat.GUI;
   /** Renderer */
   readonly renderer: THREE.WebGLRenderer;
-  /** Controls */
-  readonly controls: Controls;
 
   // Container
   container: THREE.Object3D;
@@ -40,7 +37,6 @@ export default class Camera {
     config: Config;
     debug: dat.GUI;
     renderer: THREE.WebGLRenderer;
-    controls: Controls;
   }) {
     // Options
     this.time = _params.time;
@@ -48,7 +44,6 @@ export default class Camera {
     this.config = _params.config;
     this.debug = _params.debug;
     this.renderer = _params.renderer;
-    this.controls = _params.controls;
 
     // Set up
     this.container = new THREE.Object3D();
@@ -88,6 +83,8 @@ export default class Camera {
     );
     this.cameraControls.setTarget(0, 7.5, 0);
 
+    this.cameraControls.touches.two = CameraControls.ACTION.TOUCH_ROTATE;
+
     this.sizes.on('resize', () => {
       this.instance.aspect = this.sizes.viewport.aspect;
       this.instance.updateProjectionMatrix();
@@ -102,23 +99,6 @@ export default class Camera {
         this.target.z
       );
       this.cameraControls.update(this.time.delta);
-
-      if (this.config.touch) {
-        const speed = Math.PI / 90;
-        const angle = this.controls.touch.joysticks.right.angle!.value;
-
-        if (angle === 0) {
-          return;
-        } else if (angle > 1.4 && angle < 2.8) {
-          this.cameraControls.rotate(0, -speed);
-        } else if (angle > -1.6 && angle < -0.2) {
-          this.cameraControls.rotate(0, speed);
-        } else if (angle > -0.2 && angle < 1.4) {
-          this.cameraControls.rotate(speed, 0);
-        } else {
-          this.cameraControls.rotate(-speed, 0);
-        }
-      }
     });
 
     if (this.debug) {