+++ src/typescript/World/Physics.ts
@@ -1,6 +1,5 @@
 import CANNON from 'cannon';
 import * as THREE from 'three';
-
 import Time from '../Utils/Time';
 import Controls from '../Controls';
 
@@ -103,17 +102,46 @@ export default class Physics {
         // NO OP - just forward direction.
       }
 
-      const impulse = to2dDirection(eulerDirection);
+      const impulse = to2dDirectionVector(eulerDirection).normalize();
 
       impulse.multiplyScalar(speed);
 
       this.avatar.position.add(impulse);
     };
 
+    const touchMove = (angle: number, speed: number) => {
+      if (angle === 0) return;
+
+      angle -= Math.PI / 2 - this.avatar.direction.y;
+
+      const impulse = to2dDirectionVector(new THREE.Euler(0, angle, 0));
+
+      impulse.normalize();
+      impulse.multiplyScalar(speed);
+
+      this.avatar.position.add(impulse);
+    };
+
+    const movementSpeed = 0.5;
+    const rotationSpeed = 0.1;
+    const boxCenter = new THREE.Vector3();
+    const boxToAvatarNormal = new THREE.Vector3();
     this.time.on('tick', () => {
-      const speed = 0.5;
       if (!this.avatar.boundingBox) return;
-      if (this.avatarIsColliding()) move(direction.back, undefined, speed);
+
+      // handle collisions...
+      const offendingBox = this.avatarCollidingWithBox();
+      if (offendingBox !== undefined) {
+        offendingBox.getCenter(boxCenter);
+
+        boxToAvatarNormal.subVectors(this.avatar.position, boxCenter);
+        boxToAvatarNormal.y = 0;
+        boxToAvatarNormal.normalize();
+        boxToAvatarNormal.multiplyScalar(movementSpeed * 0.2);
+
+        this.avatar.position.add(boxToAvatarNormal);
+        return;
+      }
 
       // Controls
       if (!this.config.touch) {
@@ -128,18 +156,14 @@ export default class Physics {
         else if (actions.strafeRight) strafe = direction.right;
 
         if (movement !== undefined || strafe !== undefined) {
-          move(movement, strafe, speed);
+          move(movement, strafe, movementSpeed);
         }
 
-        if (actions.left) rotate(direction.left, speed / 5);
-        if (actions.right) rotate(direction.right, speed / 5);
+        if (actions.left) rotate(direction.left, rotationSpeed);
+        if (actions.right) rotate(direction.right, rotationSpeed);
       } else {
         const angle = this.controls.touch.joysticks.left.angle!.value;
-        if (angle === 0) return;
-        else if (angle > 1.4 && angle < 2.8) this.avatar.position.z -= speed;
-        else if (angle > -1.6 && angle < -0.2) this.avatar.position.z += speed;
-        else if (angle > -0.2 && angle < 1.4) this.avatar.position.x += speed;
-        else this.avatar.position.x -= speed;
+        touchMove(angle, movementSpeed);
       }
     });
   }
@@ -151,7 +175,70 @@ export default class Physics {
    */
   setCameraControls() {
     this.time.on('tick', () => {
-      this.controls.camera.azimuthAngle = this.avatar.direction.y;
+      if (!this.config.touch) {
+        this.controls.camera.azimuthAngle = this.avatar.direction.y;
+      } else {
+        const angle =
+          this.controls.touch.joysticks.right.angle!.value - Math.PI / 2;
+
+        // For the purposes of this block of code, the right touch joystick is
+        // divided into its 4 ordinal semicircles:
+        //
+        //                    Semi Top
+        //          S        _________       S
+        //          e       /    |    \      e
+        //          m      /     |     \     m
+        //          i     /      |      \    i
+        //          L    |_______|_______|   R
+        //          e     \      |      /    i
+        //          f      \     |     /     g
+        //          t       \____|____/      h
+        //                  Semi Bottom      t
+
+        // The distance of a semicircle is the distance the joystock cursor is
+        // away from the semicircle midpoint.
+        let distSemiT = 0;
+        let distSemiL = 0;
+        let distSemiR = 0;
+        let distSemiB = 0;
+
+        // Top semicircle
+        if (Math.abs(angle) < Math.PI / 2) distSemiT = Math.abs(angle);
+        // Left semicircle
+        if (angle < Math.PI / 2 && angle > 0) {
+          distSemiL = Math.PI / 2 - angle;
+        }
+        if (angle < -Math.PI && angle > -1.5 * Math.PI) {
+          distSemiL = 1.5 * Math.PI + angle;
+        }
+        // Bottom semicircle
+        if (angle > -1.5 * Math.PI && angle < -Math.PI / 2) {
+          distSemiB = Math.abs(Math.PI + angle);
+        }
+        // Right semicircle
+        if (angle < 0 && angle > -Math.PI) {
+          distSemiR = Math.abs(Math.PI / 2 + angle);
+        }
+
+        let azimuthAngle = 0;
+        if (distSemiL > 0) azimuthAngle = Math.PI / 2 - distSemiL;
+        if (distSemiR > 0) azimuthAngle = -(Math.PI / 2 - distSemiR);
+        if (Math.abs(azimuthAngle) < Math.PI / 8) azimuthAngle = 0;
+
+        let polarAngle = 0;
+        if (distSemiT > 0) polarAngle = Math.PI / 2 - distSemiT;
+        if (distSemiB > 0) polarAngle = -(Math.PI / 2 - distSemiB);
+        if (Math.abs(polarAngle) < Math.PI / 8) polarAngle = 0;
+
+        const azimuthSpeed = 0.02;
+        const polarSpeed = 0.008;
+        this.controls.camera.rotate(
+          azimuthAngle * azimuthSpeed,
+          polarAngle * polarSpeed
+        );
+
+        this.avatar.direction.y = this.controls.camera.azimuthAngle;
+      }
     });
   }
 
@@ -171,7 +258,7 @@ export default class Physics {
   /**
    * Determines if _box is colliding with any collisionBoundingBoxes.
    */
-  avatarIsColliding() {
+  avatarCollidingWithBox(): undefined | THREE.Box3 {
     if (!this.avatar.boundingBox) {
       throw console.error('No avatar bounding box.');
     }
@@ -181,16 +268,24 @@ export default class Physics {
     const avatarMidPoint = new THREE.Vector3().copy(avatarBoundingBox.min);
     avatarMidPoint.add(avatarBoundingBox.max.clone().multiplyScalar(0.5));
 
-    return this.collisionBoundingBoxes.some(box => {
-      return box.intersectsBox(this.avatar.boundingBox!);
+    let offendingBox: THREE.Box3 | undefined = undefined;
+
+    this.collisionBoundingBoxes.some(box => {
+      if (box.intersectsBox(this.avatar.boundingBox!)) {
+        offendingBox = box.clone();
+        return true;
+      }
+      return false;
     });
+
+    return offendingBox;
   }
 }
 
 /**
  * Converts a Euler to a directional vector.
  */
-function to2dDirection(euler: THREE.Euler) {
+function to2dDirectionVector(euler: THREE.Euler) {
   // Same as finding the side lengths of a triangle with angle euler.y with
   // adjacent side as z and opposite side as x since we are in a 2D plane.
   return new THREE.Vector3(-Math.sin(euler.y), 0, -Math.cos(euler.y));